requirements.txt
flask==3.1.0
line-bot-sdk==3.11.0
requests==2.32.3
python-dotenv==1.1.0
gunicorn==23.0.0
pytz==2025.2
